pyproject.toml
[project]
name = "llm-game-rec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.13",
    "google-generativeai>=0.8.5",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "streamlit>=1.46.1",
]
